Added bounds to all editable numeric values
